fix: added logic to call create appkit before use appkit hook
